design_team/requirements.txt
google-adk
python-dotenv
sqlalchemy
aiosqlite
fastapi
uvicorn[standard]
websockets
pydantic